Store active context when pushing a closure
It is impossible to lookup a frame marker after a block was executed, but Smalltalk code does access the home context of an executed block sometimes. At the moment, it is still unclear whether the entire chain needs to be forced to heap and if the frame needs to be devirtualized.