Implement SqueakSSL Primitives (#16)
* Implement SqueakSSL primitives (#3)
* Cleanups and improvements for SqueakSSL project
* Add .mx cache to .gitignore
* Set Java compliance property to '8+'

This fixes the following behavior:
When trying to reproduce a test failure which
occurred on JDK 11 only, I set JAVA_HOME
accordingly and ran "mx unittest". However,
test discovery yielded no tests, because
no modules where found that fit the Java
version (given the previous configuration
value of '1.8').

Also, when having building with both
JDK 8 and 11 on CI, '8+' is likely the
logical compliance value that fits the
project setup best.