Print only relevant stack traces on test failure
First, replace the result of type String with a proper
data class TestResult, which can also hold a reference
to the original exception that was causing the test
failure.

Second, use the original exception, and directly
rethrow it in the test method.
This way, stack traces with the "assertFalse" line
as top-level element disappear, which presents
an improvement over "print originating stacktrace.
then print stacktrace of assertFalse(passed)".

MX only knows to stacktrace modes:
(1) print ST inline and at end of suite
(2) print ST at end of suite only
Sadly "print ST inline" only does not appear to be
viable. Therefore we are left with (1), because
it is convenient to see the test run and ST
co-located.