Move primitive execution back into bytecode loop
This ensures the bytecode loop is always reached, which in turn works around a compiler problem (see https://github.com/oracle/graal/issues/1698).

According to @timfel, callPrimitive can occur in the middle of a method's bytecode in Sista. So it's necessary to move it back into the bytecode loop anyway.