Revise send bytecodes and message dispatch
- Split self and super sends
- Introduce dispatch package
- Divide message dispatch into class lookup, selector lookup, and dispatch
- Use object layout as guard when looking up class of AbstractPointersObjects (this eliminates successive checks of the same object layout, e.g. when accessing the object)
- Use methodDictStableAssumptions from class to methodClass for selector lookups (if one of them changes, the lookup should be invalidated)

This prepares TruffleSqueak for the new inlining approach [1] shipping with GraalVM 20.2.0. Reporting of polymorphism is significantly reduced (only in DispatchLookupResultNode for now). AWFY benchmark show that overall performance is about the same with this change: While benchmarks such as Towers, Richards, and a few others benefit from the new inlining, other benchmarks show slowdowns (e.g. Havlak, List, Json, DeltaBlue). The results are worse with the previous approach for dispatching messages.

[1] https://github.com/oracle/graal/blob/91cb0529ea6540cbda7c284f56f6dad5a65d1ace/truffle/docs/Inlining.md