Show bytecode coverage in CallTargetBrowser
Since bytecode nodes are lazily inserted into Truffle AST, we can use this information to find out whether a bytecode has been executed or not. This displays this information in the bytecode view of the CallTargetBrowser. Red means not covered, green means covered, gray means unknown (closures are currently not supported).