Revert 00d1b84e30d8b6ff5fa67f00f4e441460ca94201
While only reporting megamorphic cases avoids excessive splitting in the List benchmark, this does not seem to be something TruffleSqueak needs to take care of. The List benchmark is polymorphic in Squeak/Smalltalk due to the isNil message sends, which is not the case in others (e.g. SOMns). As a result, the List benchmark is not only highly recursive but also polymorphic, which seems to be a worst case for Truffle's splitting heuristic.