Improve primitiveClosureValueNoContextSwitch
Up until now, we reused primitives 201 and 206 for 221 and 222, but that can lead to problems in many important methods (such as `ensure:`, `ifCurtailed:` and `releaseCriticalSection:`).

This commit introduces separate implementations for primitives 221 and 222 which implement the OpenSmalltalkVM behavior by delaying the next check for interrupts by 20x the time of a normal check.